ref(feedback): Move UserFeedback type into feedback.ts
