+++ packages/types/src/feedback.ts
@@ -1,6 +1,17 @@
 import type { Event } from './event';
+import type { User } from './user';
 
-export interface FeedbackContext extends Record<string, unknown> {
+/**
+ * Crash report feedback object
+ */
+export interface UserFeedback {
+  event_id: string;
+  email: User['email'];
+  name: string;
+  comments: string;
+}
+
+interface FeedbackContext extends Record<string, unknown> {
   message: string;
   contact_email?: string;
   name?: string;
