+++ packages/types/src/index.ts
@@ -69,7 +69,7 @@ export type {
   Profile,
 } from './profiling';
 export type { ReplayEvent, ReplayRecordingData, ReplayRecordingMode } from './replay';
-export type { FeedbackEvent } from './feedback';
+export type { FeedbackEvent, UserFeedback } from './feedback';
 export type { QueryParams, Request, SanitizedRequestData } from './request';
 export type { Runtime } from './runtime';
 export type { CaptureContext, Scope, ScopeContext, ScopeData } from './scope';
@@ -131,7 +131,7 @@ export type {
   BaseTransportOptions,
   TransportRequestExecutor,
 } from './transport';
-export type { User, UserFeedback } from './user';
+export type { User } from './user';
 export type { WebFetchHeaders, WebFetchRequest } from './webfetchapi';
 export type { WrappedFunction } from './wrappedfunction';
 export type {
