+++ packages/types/src/user.ts
@@ -8,10 +8,3 @@ export interface User {
   email?: string;
   username?: string;
 }
-
-export interface UserFeedback {
-  event_id: string;
-  email: User['email'];
-  name: string;
-  comments: string;
-}
